apply updates
